+++ src/NetStandard/E.Standard.Web/Extensions/RequestMessageExtensions.cs
@@ -30,13 +30,13 @@ static public class RequestMessageExtensions
 
             if (!String.IsNullOrEmpty(username) &&
                 !String.IsNullOrEmpty(password) &&
-                authType?.ToLower() == "basic")
+                authType?.Equals("basic", StringComparison.OrdinalIgnoreCase) == true)
             {
                 requestMessage.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Basic",
                     Convert.ToBase64String(Encoding.UTF8.GetBytes($"{username}:{password}")));
             }
             else if (!String.IsNullOrEmpty(authentication.AccessToken) &&
-                    authType?.ToLower() == "bearer")
+                    authType?.Equals("bearer", StringComparison.OrdinalIgnoreCase) == true)
             {
                 requestMessage.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", authentication.AccessToken);
             }